Update dataset.py
Using torchaudio instead of librosa to contour conflicts in C++ on Umiacs.

Change filename reading.

TODO: The video read is of shape 3*224*224, but should be of shape 3*2*224*224 since 2 frames per clip.

Spectrogram need to be regenerated directly from audio, since the image is saved with axis, and the original implementation reads data of shape 4*400*1000. I have no idea why it is 4 channels.